android_port/requirements.txt
kivy==2.1.0
kivymd
plyer
pyjnius
android
